remove type conversion in CGCNN.forward, which was causing tensor allocation on CPU during GPU training. (alternative would be to specify device or typecast in place, but it seems cleaner to pass consistent input types

+++ jarvisdgl/data.py
@@ -47,7 +47,9 @@ def dgl_crystal(
     g.edata["bondlength"] = g.edata["weight"].type(torch.FloatTensor)
     del g.edata["weight"]
 
-    g.ndata["atom_features"] = torch.tensor(jgraph.node_attributes)
+    g.ndata["atom_features"] = torch.tensor(jgraph.node_attributes).type(
+        torch.FloatTensor
+    )
 
     return g
